+++ app/src/main/java/com/duckduckgo/app/browser/remotemessage/CommandActionMapper.kt
@@ -16,8 +16,8 @@
 
 package com.duckduckgo.app.browser.remotemessage
 
-import com.duckduckgo.app.browser.BrowserTabViewModel.Command
-import com.duckduckgo.app.browser.BrowserTabViewModel.Command.*
+import com.duckduckgo.app.browser.commands.Command
+import com.duckduckgo.app.browser.commands.Command.*
 import com.duckduckgo.remote.messaging.api.Action
 import com.duckduckgo.remote.messaging.api.Action.*
